Free version update 1-8-2024.

+++ L2J_Mobius_11.1_TheSourceOfFlame/java/org/l2jmobius/gameserver/network/clientpackets/RequestMultisellList.java
@@ -31,7 +31,7 @@ public class RequestMultisellList extends ClientPacket
 	@Override
 	protected void readImpl()
 	{
-		_multiSellId = readShort();
+		_multiSellId = readInt();
 	}
 	
 	@Override
